Created SafeWinchingClimber
This class will manage a pneumatic arm and single-use winch in order to
eliminate as much risk of user error as possible. The arm can be raised
and lowered as much as necessary, and the winch can only be activated
when the arm is in top position (that requirement only works if the time
the arm takes to extend is provided in the constructor). Two inputs are
required to trigger the winch, so accidental activation is mostly
eliminated. Once the winch begins, the pneumatic arm is deactivated and
disabled.