SafeWinchingClimber will now retract arm correctly on winch.
It was discovered that DoubleSolenoid.Value.kOff does not actually
disable pressure in the arm. It only cuts off current flowing to the
arm's solenoid. This will not unpressurize it but leave it in the
current state. That would cause problems during winching since it would
compete directly with the motor. Now, the arm will retract during
winching, so it will assist the motor in lifting the robot.